Provide caveat around critical applications
I have trialled replacing a hand coded application driving relays, driven with MQTT, with a ESPHome setup for a WEMOS UNO style board and a quad relay shield.  Works a treat under automations driven by HA if the server behaves.  If server crashes or wifi drops out between server and ESPHome widget, the ESPHome widget turned on 3 of the 4 relays.  The problem does not clear if the server or wifi comes back up, so the widget seems to have crashed and does not re-connect.  The automations in HA are one per switch and run on/off cycles with time gaps to avoid the sprinkler system sharing water pressure between zones.  The ESPHome widget is on a test bench so is not connected to solenoids so there are no opportunities for spurious pulses due to solenoids, there is TVS in circuit in any event.   HA does report the ESPHome widget in the state it is in with 3 of 4 switches on.  The automations however, as reported in the log, report all widgets off.  Whether this behaviour is in HA or ESPHome does not matter, what matters is on when off intended or off when on intended is a problem for any system.  So watering, motors being controlled, anything with physical energy or material flow will be a problem when this condition arises.   While there is likely a general caveat about use of the software buried where I can't currently find it, it might be best to repeat declarations throughout.  A page about the problems of critical applications and how combining different solutions from different providers and managing states and modes and who need wear the target in court might be apt.